+++ docs/scripts/views/menu.vue
@@ -31,6 +31,8 @@
       </div>
     </div>
     <ui-markdown :text="code[1]"></ui-markdown>
+
+    // TODO
   </div>
 </template>